+++ app/src/main/java/org/iflab/wecenterandroid/modal/home/Home501.java
@@ -23,7 +23,7 @@ public class Home501 implements Home{
 
     private int history_id;
     private int associate_action;
-    private int add_time;
+    private long add_time;
     /**
      * uid : 3
      * user_name : BugFree
@@ -82,7 +82,7 @@ public class Home501 implements Home{
         this.associate_action = associate_action;
     }
 
-    public void setAdd_time(int add_time) {
+    public void setAdd_time(long add_time) {
         this.add_time = add_time;
     }
 
@@ -104,7 +104,7 @@ public class Home501 implements Home{
     }
 
     @Override
-    public int getAdd_time() {
+    public long getAdd_time() {
         return add_time;
     }
